dedicated class files
