+++ INPTPZ1/Program.cs
@@ -126,139 +126,4 @@ namespace INPTPZ1
             bmp.SetPixel(j, i, vv);
         }
     }
-
-    class Polynome
-    {
-        public List<ComplexNumber> ComplexNumbers { get; set; }
-
-        public Polynome()
-        {
-            ComplexNumbers = new List<ComplexNumber>();
-        }
-
-        public Polynome(int count)
-        {
-            ComplexNumbers = new List<ComplexNumber>
-            {
-                new ComplexNumber() { RealPart = 1 }
-            };
-
-            for (int i = 0; i < count - 1; i++)
-            {
-                ComplexNumbers.Add(ComplexNumber.Zero);
-            }
-            ComplexNumbers.Add(new ComplexNumber() { RealPart = 1 });
-        }
-
-        public Polynome Derive()
-        {
-            Polynome p = new Polynome();
-
-            for (int i = 1; i < ComplexNumbers.Count; i++)
-            {
-                p.ComplexNumbers.Add(ComplexNumbers[i].Multiply(new ComplexNumber() { RealPart = i }));
-            }
-
-            return p;
-        }
-
-        public ComplexNumber Evaluate(ComplexNumber x)
-        {
-            ComplexNumber s = ComplexNumber.Zero;
-
-            for (int i = 0; i < ComplexNumbers.Count; i++)
-            {
-                ComplexNumber complexNumber = ComplexNumbers[i];
-                ComplexNumber complexNumberEvaluated = x;
-                int power = i;
-
-                if (i > 0)
-                {
-                    for (int j = 0; j < power - 1; j++)
-                        complexNumberEvaluated = complexNumberEvaluated.Multiply(x);
-                    complexNumber = complexNumber.Multiply(complexNumberEvaluated);
-                }
-                s = s.Add(complexNumber);
-            }
-
-            return s;
-        }
-
-        public override string ToString()
-        {
-            string output = "";
-
-            for (int i = 0; i < ComplexNumbers.Count; i++)
-            {
-                output += ComplexNumbers[i];
-                if (i > 0)
-                {
-                    for (int j = 0; j < i; j++)
-                        output += "x";
-                }
-                output += " + ";
-            }
-
-            return output;
-        }
-    }
-
-    class ComplexNumber
-    {
-        public double RealPart { get; set; }
-        public float ImaginaryPart { get; set; }
-
-        public readonly static ComplexNumber Zero = new ComplexNumber()
-        {
-            RealPart = 0,
-            ImaginaryPart = 0
-        };
-
-        public ComplexNumber Multiply(ComplexNumber násobitel)
-        {
-            // aRe*bRe + aRe*bIm*i + aIm*bRe*i - aIm*bIm*i*i
-            return new ComplexNumber()
-            {
-                RealPart = this.RealPart * násobitel.RealPart - this.ImaginaryPart * násobitel.ImaginaryPart,
-                ImaginaryPart = (float)(this.RealPart * násobitel.ImaginaryPart + this.ImaginaryPart * násobitel.RealPart)
-            };
-        }
-
-        public ComplexNumber Add(ComplexNumber summand)
-        {
-            return new ComplexNumber()
-            {
-                RealPart = this.RealPart + summand.RealPart,
-                ImaginaryPart = this.ImaginaryPart + summand.ImaginaryPart
-            };
-        }
-        public ComplexNumber Subtract(ComplexNumber subtrahend)
-        {
-            return new ComplexNumber()
-            {
-                RealPart = this.RealPart - subtrahend.RealPart,
-                ImaginaryPart = this.ImaginaryPart - subtrahend.ImaginaryPart
-            };
-        }
-
-        public override string ToString()
-        {
-            return $"({RealPart} + {ImaginaryPart}i)";
-        }
-
-        internal ComplexNumber Divide(ComplexNumber divisor)
-        {
-            // (aRe + aIm*i) / (bRe + bIm*i)
-            // ((aRe + aIm*i) * (bRe - bIm*i)) / ((bRe + bIm*i) * (bRe - bIm*i))
-            //  bRe*bRe + bIm*bIm*i*i
-            ComplexNumber thisNegated = this.Multiply(new ComplexNumber() { RealPart = divisor.RealPart, ImaginaryPart = -divisor.ImaginaryPart });
-            double divideWith = divisor.RealPart * divisor.RealPart + divisor.ImaginaryPart * divisor.ImaginaryPart;
-
-            return new ComplexNumber()
-            {
-                RealPart = thisNegated.RealPart / divideWith,
-                ImaginaryPart = (float)(thisNegated.ImaginaryPart / divideWith)
-            };
-        }
-    }
 }
